ores value fix

+++ Assets/Scripts/ScoreManager.cs
@@ -7,14 +7,14 @@ public class ScoreManager : MonoBehaviour
 
     [SerializeField] private Text uitext;
 
-    public int score = 0;
-    public void AddScore(int scoreAmount)
+    public float score = 0;
+    public void AddScore(float scoreAmount)
     {
         score += scoreAmount;
         uitext.text = $"Score: {score.ToString()} ";
     }
 
-    public void RemoveScore(int scoreAmount)
+    public void RemoveScore(float scoreAmount)
     {
             score -= scoreAmount;
             uitext.text = $"Score: {score.ToString()} ";
